move pre-go1.19 wrappers to bbig/bridge
